Feat: 카드들 컴포넌트화 + mock데이터 분리 #29
