+++ src/components/MyPage/Activities/ActivitiesMock.ts
@@ -1,0 +1,257 @@
+export const myActivitiesMock = {
+  activities: [
+    {
+      id: 1,
+      userId: 10,
+      title: '함께 배우면 신나는 종민 댄스',
+      description: '둠칫 둠칫 두둠칫',
+      category: '투어',
+      price: 10000,
+      address: '서울특별시 강남구 테헤란로 427',
+      bannerImageUrl:
+        'https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/globalnomad/activity_registration_image/a.png',
+      rating: 1,
+      reviewCount: 10,
+      createdAt: '2024-01-27T11:10:44.283Z',
+    },
+    {
+      id: 2,
+      userId: 9,
+      title: '함께 배우면 즐거운 스트릿댄스',
+      description: '둠칫 둠칫 두둠칫',
+      category: '투어',
+      price: 15000,
+      address: '서울특별시 강남구 테헤란로 427',
+      bannerImageUrl:
+        'https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/globalnomad/activity_registration_image/a.png',
+      rating: 3,
+      reviewCount: 30,
+      createdAt: '2024-01-27T00:16:43.256Z',
+    },
+    {
+      id: 3,
+      userId: 8,
+      title: '함께 배우면 슬픈 스트릿댄스',
+      description: '둠칫 둠칫 두둠칫',
+      category: '투어',
+      price: 20000,
+      address: '서울특별시 강남구 테헤란로 427',
+      bannerImageUrl:
+        'https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/globalnomad/activity_registration_image/a.png',
+      rating: 5,
+      reviewCount: 50,
+      createdAt: '2024-01-27T00:16:43.256Z',
+    },
+    {
+      id: 4,
+      userId: 8,
+      title: '함께 배우면 슬픈 스트릿댄스',
+      description: '둠칫 둠칫 두둠칫',
+      category: '투어',
+      price: 20000,
+      address: '서울특별시 강남구 테헤란로 427',
+      bannerImageUrl:
+        'https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/globalnomad/activity_registration_image/a.png',
+      rating: 5,
+      reviewCount: 50,
+      createdAt: '2024-01-27T00:16:43.256Z',
+    },
+    {
+      id: 5,
+      userId: 8,
+      title: '함께 배우면 슬픈 스트릿댄스',
+      description: '둠칫 둠칫 두둠칫',
+      category: '투어',
+      price: 20000,
+      address: '서울특별시 강남구 테헤란로 427',
+      bannerImageUrl:
+        'https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/globalnomad/activity_registration_image/a.png',
+      rating: 5,
+      reviewCount: 50,
+      createdAt: '2024-01-27T00:16:43.256Z',
+    },
+    {
+      id: 6,
+      userId: 8,
+      title: '함께 배우면 슬픈 스트릿댄스',
+      description: '둠칫 둠칫 두둠칫',
+      category: '투어',
+      price: 20000,
+      address: '서울특별시 강남구 테헤란로 427',
+      bannerImageUrl:
+        'https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/globalnomad/activity_registration_image/a.png',
+      rating: 5,
+      reviewCount: 50,
+      createdAt: '2024-01-27T00:16:43.256Z',
+    },
+    {
+      id: 7,
+      userId: 8,
+      title: '함께 배우면 슬픈 스트릿댄스',
+      description: '둠칫 둠칫 두둠칫',
+      category: '투어',
+      price: 20000,
+      address: '서울특별시 강남구 테헤란로 427',
+      bannerImageUrl:
+        'https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/globalnomad/activity_registration_image/a.png',
+      rating: 5,
+      reviewCount: 50,
+      createdAt: '2024-01-27T00:16:43.256Z',
+    },
+    {
+      id: 8,
+      userId: 8,
+      title: '함께 배우면 슬픈 스트릿댄스',
+      description: '둠칫 둠칫 두둠칫',
+      category: '투어',
+      price: 20000,
+      address: '서울특별시 강남구 테헤란로 427',
+      bannerImageUrl:
+        'https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/globalnomad/activity_registration_image/a.png',
+      rating: 5,
+      reviewCount: 50,
+      createdAt: '2024-01-27T00:16:43.256Z',
+    },
+    {
+      id: 9,
+      userId: 8,
+      title: '함께 배우면 슬픈 스트릿댄스',
+      description: '둠칫 둠칫 두둠칫',
+      category: '투어',
+      price: 20000,
+      address: '서울특별시 강남구 테헤란로 427',
+      bannerImageUrl:
+        'https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/globalnomad/activity_registration_image/a.png',
+      rating: 5,
+      reviewCount: 50,
+      createdAt: '2024-01-27T00:16:43.256Z',
+    },
+    {
+      id: 10,
+      userId: 8,
+      title: '함께 배우면 슬픈 스트릿댄스',
+      description: '둠칫 둠칫 두둠칫',
+      category: '투어',
+      price: 20000,
+      address: '서울특별시 강남구 테헤란로 427',
+      bannerImageUrl:
+        'https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/globalnomad/activity_registration_image/a.png',
+      rating: 5,
+      reviewCount: 50,
+      createdAt: '2024-01-27T00:16:43.256Z',
+    },
+    {
+      id: 11,
+      userId: 8,
+      title: '함께 배우면 슬픈 스트릿댄스',
+      description: '둠칫 둠칫 두둠칫',
+      category: '투어',
+      price: 20000,
+      address: '서울특별시 강남구 테헤란로 427',
+      bannerImageUrl:
+        'https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/globalnomad/activity_registration_image/a.png',
+      rating: 5,
+      reviewCount: 50,
+      createdAt: '2024-01-27T00:16:43.256Z',
+    },
+    {
+      id: 12,
+      userId: 8,
+      title: '함께 배우면 슬픈 스트릿댄스',
+      description: '둠칫 둠칫 두둠칫',
+      category: '투어',
+      price: 20000,
+      address: '서울특별시 강남구 테헤란로 427',
+      bannerImageUrl:
+        'https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/globalnomad/activity_registration_image/a.png',
+      rating: 5,
+      reviewCount: 50,
+      createdAt: '2024-01-27T00:16:43.256Z',
+    },
+    {
+      id: 13,
+      userId: 8,
+      title: '함께 배우면 슬픈 스트릿댄스',
+      description: '둠칫 둠칫 두둠칫',
+      category: '투어',
+      price: 20000,
+      address: '서울특별시 강남구 테헤란로 427',
+      bannerImageUrl:
+        'https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/globalnomad/activity_registration_image/a.png',
+      rating: 5,
+      reviewCount: 50,
+      createdAt: '2024-01-27T00:16:43.256Z',
+    },
+    {
+      id: 14,
+      userId: 8,
+      title: '함께 배우면 슬픈 스트릿댄스',
+      description: '둠칫 둠칫 두둠칫',
+      category: '투어',
+      price: 20000,
+      address: '서울특별시 강남구 테헤란로 427',
+      bannerImageUrl:
+        'https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/globalnomad/activity_registration_image/a.png',
+      rating: 5,
+      reviewCount: 50,
+      createdAt: '2024-01-27T00:16:43.256Z',
+    },
+    {
+      id: 15,
+      userId: 8,
+      title: '함께 배우면 슬픈 스트릿댄스',
+      description: '둠칫 둠칫 두둠칫',
+      category: '투어',
+      price: 20000,
+      address: '서울특별시 강남구 테헤란로 427',
+      bannerImageUrl:
+        'https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/globalnomad/activity_registration_image/a.png',
+      rating: 5,
+      reviewCount: 50,
+      createdAt: '2024-01-27T00:16:43.256Z',
+    },
+    {
+      id: 16,
+      userId: 8,
+      title: '함께 배우면 슬픈 스트릿댄스',
+      description: '둠칫 둠칫 두둠칫',
+      category: '투어',
+      price: 20000,
+      address: '서울특별시 강남구 테헤란로 427',
+      bannerImageUrl:
+        'https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/globalnomad/activity_registration_image/a.png',
+      rating: 5,
+      reviewCount: 50,
+      createdAt: '2024-01-27T00:16:43.256Z',
+    },
+    {
+      id: 17,
+      userId: 8,
+      title: '함께 배우면 슬픈 스트릿댄스',
+      description: '둠칫 둠칫 두둠칫',
+      category: '투어',
+      price: 20000,
+      address: '서울특별시 강남구 테헤란로 427',
+      bannerImageUrl:
+        'https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/globalnomad/activity_registration_image/a.png',
+      rating: 5,
+      reviewCount: 50,
+      createdAt: '2024-01-27T00:16:43.256Z',
+    },
+    {
+      id: 18,
+      userId: 8,
+      title: '함께 배우면 슬픈 스트릿댄스',
+      description: '둠칫 둠칫 두둠칫',
+      category: '투어',
+      price: 20000,
+      address: '서울특별시 강남구 테헤란로 427',
+      bannerImageUrl:
+        'https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/globalnomad/activity_registration_image/a.png',
+      rating: 5,
+      reviewCount: 50,
+      createdAt: '2024-01-27T00:16:43.256Z',
+    },
+  ],
+  totalCount: 18,
+};
